+++ app/models/player.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, Integer, String, Float, Boolean, Date, ForeignKey
+from sqlalchemy import Column, Integer, String, Boolean, Date, ForeignKey
 from sqlalchemy.orm import relationship
 from app.database.declarative_base import Base
 
@@ -15,29 +15,34 @@ class Player(Base):
     display_last_comma_first = Column(String, nullable=True)
     display_fi_last = Column(String, nullable=True)
     player_slug = Column(String, nullable=True)
+
     birthdate = Column(Date, nullable=True)
     school = Column(String, nullable=True)
     country = Column(String, nullable=True)
     last_affiliation = Column(String, nullable=True)
     height = Column(String, nullable=True)
-    weight = Column(Float, nullable=True)
+    weight = Column(Integer, nullable=True)
     season_exp = Column(Integer, nullable=True)
     jersey = Column(Integer, nullable=True)
     position = Column(String, nullable=True)
     roster_status = Column(String, nullable=True)
+
     team_id = Column(Integer, ForeignKey("team.team_id"), nullable=True)
     team_name = Column(String, nullable=True)
     team_abbreviation = Column(String, nullable=True)
     team_code = Column(String, nullable=True)
     team_city = Column(String, nullable=True)
+
     player_code = Column(String, nullable=True)
     from_year = Column(Integer, nullable=True)
     to_year = Column(Integer, nullable=True)
+
     dleague_flag = Column(Boolean, nullable=True)
     nba_flag = Column(Boolean, nullable=True)
     games_played_flag = Column(Boolean, nullable=True)
     games_played_current_season_flag  = Column(Boolean, nullable=True)
     greatest_75_flag = Column(Boolean, nullable=True)
+
     draft_year = Column(String, nullable=True)      # String car peut être Undrafted
     draft_round = Column(String, nullable=True)     # String car peut être Undrafted
     draft_number = Column(String, nullable=True)    # String car peut être Undrafted
